+++ src/Recipe/RecipeInterface.php
@@ -40,12 +40,4 @@ use Teknoo\Recipe\RecipeInterface as BaseInterface;
  */
 interface RecipeInterface extends BaseInterface
 {
-    /*
-     * Method to register middleware in the manager to process request.
-     */
-    public function registerMiddleware(
-        MiddlewareInterface $middleware,
-        int $priority = 10,
-        ?string $middlewareName = null
-    ): RecipeInterface;
 }