+++ tests/universal/Http/Bowl/PSR15/HandlerBowlTest.php
@@ -96,7 +96,7 @@ class HandlerBowlTest extends AbstractBowlTests
         );
     }
 
-    public function testExecute()
+    public function testExecute(): void
     {
         $values = $this->getValidWorkPlan();
         self::assertInstanceOf(
@@ -109,7 +109,7 @@ class HandlerBowlTest extends AbstractBowlTests
         );
     }
 
-    public function testExecuteWithValue()
+    public function testExecuteWithValue(): void
     {
         self::assertTrue(true);
     }